Fix join table for many-to-many relation and displaying history of orders, add migration

+++ HappyGift/HappyGift/Models/GiftServices.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -10,10 +11,11 @@ namespace HappyGift.Models
     {
         [Key]
         public int GiftServiceId { get; set; }
-        public int GiftId { get; set; }
+
+        public long GiftId { get; set; }
         public virtual Gift Gift{get; set;}
 
-        public int ServiceId { get; set; }
+        public long ServiceId { get; set; }
         public virtual Service Service { get; set; }
     }
 }
